quizDatabase.java 1.0
created java class file for operating with the quizzes in the database. it has only add quiz function in it and im working with the findQuiz function and it's yet to be added. also changed createQuiz servlet which now adds user submitted quiz straight to the database and not in servletContext.